Add custom styles, enhance audio service with VoiceSettings, and create initial HTML structure
