+++ frontend/style.css
@@ -1,0 +1,18 @@
+/* Custom overrides if any */
+.prose p {
+    margin-bottom: 1rem;
+  }
+  
+  .prose h2 {
+    margin-top: 1.5rem;
+    margin-bottom: 1rem;
+  }
+  
+  .prose blockquote {
+    border-left: 4px solid #ccc;
+    padding-left: 1rem;
+    font-style: italic;
+  }
+  
+  /* Add any additional custom styling as needed */
+
